requirements.txt
# AWS Strands SDK
strands-agents>=0.1.0
strands-tools>=0.1.0

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# VCS Integration
PyGithub>=2.1.1
python-gitlab>=4.2.0
GitPython>=3.1.40

# Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.1
httpx>=0.25.2
aiofiles>=23.2.1

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
